+++ app/admin/data_uploads.rb
@@ -68,17 +68,14 @@ ActiveAdmin.register DataUpload do
 
   controller do
     def create
-      @data_upload = build_new_resource
+      puts permitted_params[:data_upload].inspect
+      @data_upload = Command::CsvDataUpload.new(permitted_params[:data_upload])
 
-      if @data_upload.save
+      if @data_upload.call
         redirect_to admin_data_upload_path(@data_upload.data_upload)
       else
         render :new
       end
-    end
-
-    def build_new_resource
-      DataUploadForm.new(permitted_params[:data_upload])
     end
   end
 end
