+++ spec/support/fixtures/files/legislations_simple.csv
@@ -1,0 +1,3 @@
+"Id","Law","Title","Date passed","Description","Geography","Geography iso","Frameworks","Document types","Visibility status"
+"10","101","Finance Act 2011","01 Jan 2012","Description","United Kingdom","GBR","","Law","draft"
+"20","102","Climate Law","15 Jan 2015","Description","Poland","PL","","Law","pending"
