+++ spec/support/fixtures/files/litigations_simple.csv
@@ -1,0 +1,5 @@
+"Id","Title","Document type","Geography iso","Geography","Jurisdiction iso","Jurisdiction","Citation reference number","Summary","Keywords","Core objective","Visibility status","Legislation ids"
+"125","R. on the application of Tate & Lyle Industries Ltd v. Secretary of State for Energy","Case","GBR","United Kingdom","GBR","United Kingdom","[2010] EWHC 2752","Lyle requested judicial review ..","","","Draft",""
+"103","In the Matter of An Application by Brian Quinn and Michael Quinn","Case","GBR","United Kingdom","GBR","United Kingdom","[2013] NIQB 24","The applicants were brothers ...","","","Draft",""
+"93","R. on the application of Corbett v. Cornwall Council","Case","GBR","United Kingdom","GBR","United Kingdom","[2013] EWHC 3958","Local council granted approval ...","","Challenge the permit approval for the construction of a wind farm","Draft",""
+"92","R. on the application of Griffin v. Newham London Borough Council","Case","GBR","United Kingdom","GBR","United Kingdom","[2011] EWHC 53","Fight the Fights, ..","","To challenge local council decision granting consent to increased flights","Draft",""
